Extract filters component spec helpers
